use getServerSideProps Rename getStaticProps to getServerSideProps.
