+++ packages/app-content-pages/pages/about/publications.js
@@ -1,1 +1,1 @@
-export { default, getStaticProps } from '../../src/screens/Publications'
+export { default, getServerSideProps } from '../../src/screens/Publications'
