+++ packages/app-content-pages/src/screens/Team/getServerSideProps.js
@@ -1,7 +1,7 @@
 import createTeamResponse  from '../../api/team'
 import cache from '../../api/team/teamCache'
 
-export default async function getStaticProps() {
+export default async function getServerSideProps() {
   let error = null
   let teamData = []
   try {
