+++ packages/app-content-pages/src/screens/Team/getServerSideProps.spec.js
@@ -1,12 +1,12 @@
 import sinon from 'sinon'
 
 import cache from '../../api/team/teamCache'
-import getStaticProps from './getStaticProps'
+import getServerSideProps from './getServerSideProps'
 import mockData from './TeamContainer.mock'
 
 const DATA = mockData
 
-describe('Component > TeamContainer > getStaticProps', function () {
+describe('Component > TeamContainer > getServerSideProps', function () {
 
   describe('populates the "teamData" props from contentful API', function () {
     let getTeamDataStub
@@ -17,7 +17,7 @@ describe('Component > TeamContainer > getStaticProps', function () {
 
     it('should handle valid API data', async () => {
       getTeamDataStub = sinon.stub(cache, 'get').returns(DATA)
-      const { props } = await getStaticProps({})
+      const { props } = await getServerSideProps({})
       expect(props).to.deep.equal({
         error: null,
         teamData: DATA
@@ -26,7 +26,7 @@ describe('Component > TeamContainer > getStaticProps', function () {
 
     it('should handle empty API reponse', async () => {
       getTeamDataStub = sinon.stub(cache, 'get').returns([])
-      const { props } = await getStaticProps({})
+      const { props } = await getServerSideProps({})
       expect(props).to.deep.equal({
         error: null,
         teamData: []
@@ -37,7 +37,7 @@ describe('Component > TeamContainer > getStaticProps', function () {
       var errorMsg = 'failed to connect to API'
       var errorPromise = Promise.reject(new Error(errorMsg))
       getTeamDataStub = sinon.stub(cache, 'get').returns(errorPromise)
-      const { props } = await getStaticProps({})
+      const { props } = await getServerSideProps({})
       expect(props).to.deep.equal({
         error: errorMsg,
         teamData: []
