+++ packages/app-content-pages/src/screens/Team/index.js
@@ -1,3 +1,3 @@
 export { default } from './TeamContainer'
-export { default as getStaticProps } from './getStaticProps'
+export { default as getServerSideProps } from './getServerSideProps'
